requirements.txt
# Core
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Database
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy==2.0.23

# Excel & PDF Generation
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
pypdf2==3.0.1

# Charts & Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Email & Distribution
sendgrid==6.11.0
python-smtp==1.0.0

# Task Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Utilities
click==8.1.7
tqdm==4.66.1
colorlog==6.8.0
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
